Create app js to hold all files and move login to new component
